♻️ refactor: 아이템 디테일 컴포넌트 분리
