+++ src/components/items/ItemCard.tsx
