+++ src/constants/massages.ts
@@ -1,0 +1,1 @@
+export const ITEM_NOT_FOUND_MESSAGE = "아이템을 찾을 수 없습니다.";
